+++ .gitignore
@@ -3,5 +3,7 @@ venv/
 src/__pycache__/
 src/**/__pycache__/
 tests/__pycache__/
+tests/**/__pycache__/
 .vscode/
 movie_script_generator.log
+*.egg-info
