+++ src/client/__init__.py
@@ -1,0 +1,10 @@
+"""
+A2A Protocol Client Library
+This module provides a client implementation for the A2A protocol.
+"""
+
+from .agent_client import AgentClient
+from .agent_interpreter import AgentCardInterpreter
+from .models import TaskStatus, TaskArtifact
+
+__all__ = ['AgentClient', 'AgentCardInterpreter', 'TaskStatus', 'TaskArtifact']
