+++ tests/e2e/__init__.py
@@ -1,0 +1,3 @@
+"""
+End-to-end tests for the A2A protocol implementation
+"""
